+++ generators/app/templates/blank/js/main.js
@@ -9,58 +9,56 @@
  */
 
 requirejs.config(
-{
-  baseUrl: 'js',
+  {
+    baseUrl: 'js',
 
-  // Path mappings for the logical module names
-  // Update the main-release-paths.json for release mode when updating the mappings
-  paths:
-  //injector:mainReleasePaths
-  {
-    'knockout': 'libs/knockout/knockout-3.4.0.debug',
-    'jquery': 'libs/jquery/jquery-3.1.1',
-    'jqueryui-amd': 'libs/jquery/jqueryui-amd-1.12.0',
-    'promise': 'libs/es6-promise/es6-promise',
-    'hammerjs': 'libs/hammer/hammer-2.0.8',
-    'ojdnd': 'libs/dnd-polyfill/dnd-polyfill-1.0.0',
-    'ojs': 'libs/oj/v3.2.0/debug',
-    'ojL10n': 'libs/oj/v3.2.0/ojL10n',
-    'ojtranslations': 'libs/oj/v3.2.0/resources',
-    'text': 'libs/require/text',
-    'signals': 'libs/js-signals/signals',
-    'customElements': 'libs/webcomponents/CustomElements',
-    'proj4': 'libs/proj4js/dist/proj4-src',
-    'css': 'libs/require-css/css'
-  }
-  //endinjector
-  ,
-  // Shim configurations for modules that do not expose AMD
-  shim:
-  {
-    'jquery':
+    // Path mappings for the logical module names
+    // Update the main-release-paths.json for release mode when updating the mappings
+    paths:
+    //injector:mainReleasePaths
     {
-      exports: ['jQuery', '$']
+      'knockout': 'libs/knockout/knockout-3.4.0.debug',
+      'jquery': 'libs/jquery/jquery-3.1.1',
+      'jqueryui-amd': 'libs/jquery/jqueryui-amd-1.12.0',
+      'promise': 'libs/es6-promise/es6-promise',
+      'hammerjs': 'libs/hammer/hammer-2.0.8',
+      'ojdnd': 'libs/dnd-polyfill/dnd-polyfill-1.0.0',
+      'ojs': 'libs/oj/v4.0.0/debug',
+      'ojL10n': 'libs/oj/v4.0.0/ojL10n',
+      'ojtranslations': 'libs/oj/v4.0.0/resources',
+      'text': 'libs/require/text',
+      'signals': 'libs/js-signals/signals',
+      'customElements': 'libs/webcomponents/custom-elements.min',
+      'proj4': 'libs/proj4js/dist/proj4-src',
+      'css': 'libs/require-css/css'
+    }
+    //endinjector
+    ,
+
+    // Shim configurations for modules that do not expose AMD
+    shim:
+    {
+      'jquery':
+      {
+        exports: ['jQuery', '$']
+      }
     }
   }
-}
 );
 
 require(['ojs/ojcore', 'knockout', 'jquery', 'ojs/ojknockout'],
-  function(oj, ko, $)
-  {
-    $(function()
-    {      
+  function (oj, ko, $) {
+    $(function () {
       function init() {
       }
-      
-      // If running in a hybrid (e.g. Cordova) environment, we need to wait for the deviceready 
+
+      // If running in a hybrid (e.g. Cordova) environment, we need to wait for the deviceready
       // event before executing any code that might interact with Cordova APIs or plugins.
       if ($(document.body).hasClass('oj-hybrid')) {
-        document.addEventListener("deviceready", init);
+        document.addEventListener('deviceready', init);
       } else {
         init();
       }
-      
     });
   }
 );