+++ generators/hybrid/templates/blank/js/main.js
@@ -12,10 +12,10 @@ requirejs.config(
   {
     baseUrl: 'js',
 
-  // Path mappings for the logical module names
-  // Update the main-release-paths.json for release mode when updating the mappings
+    // Path mappings for the logical module names
+    // Update the main-release-paths.json for release mode when updating the mappings
     paths:
-  //injector:mainReleasePaths
+    //injector:mainReleasePaths
     {
       'knockout': 'libs/knockout/knockout-3.4.0.debug',
       'jquery': 'libs/jquery/jquery-3.1.1',
@@ -23,18 +23,18 @@ requirejs.config(
       'promise': 'libs/es6-promise/es6-promise',
       'hammerjs': 'libs/hammer/hammer-2.0.8',
       'ojdnd': 'libs/dnd-polyfill/dnd-polyfill-1.0.0',
-      'ojs': 'libs/oj/v3.2.0/debug',
-      'ojL10n': 'libs/oj/v3.2.0/ojL10n',
-      'ojtranslations': 'libs/oj/v3.2.0/resources',
+      'ojs': 'libs/oj/v4.0.0/debug',
+      'ojL10n': 'libs/oj/v4.0.0/ojL10n',
+      'ojtranslations': 'libs/oj/v4.0.0/resources',
       'text': 'libs/require/text',
       'signals': 'libs/js-signals/signals',
-      'customElements': 'libs/webcomponents/CustomElements',
+      'customElements': 'libs/webcomponents/custom-elements.min',
       'proj4': 'libs/proj4js/dist/proj4-src',
       'css': 'libs/require-css/css'
     }
     //endinjector
     ,
-  // Shim configurations for modules that do not expose AMD
+    // Shim configurations for modules that do not expose AMD
     shim:
     {
       jquery:
@@ -46,10 +46,8 @@ requirejs.config(
 );
 
 require(['ojs/ojcore', 'knockout', 'jquery', 'ojs/ojknockout'],
-  function(oj, ko, $)
-  {
-    $(function() 
-    {
+  function (oj, ko, $) {
+    $(function () {
       function init() {
       }