+++ util/paths.js
@@ -17,7 +17,8 @@ module.exports =
   getConfiguredPaths(appDir) {
     let resultPaths = _cachedPathsMap[appDir];
     if (!resultPaths) {
-      _cachedPathsMap[appDir] = resultPaths = _deriveConfiguredPaths(appDir);
+      resultPaths = _deriveConfiguredPaths(appDir);
+      _cachedPathsMap[appDir] = resultPaths;
     }
     return resultPaths;
   },
@@ -34,6 +35,7 @@ function _deriveConfiguredPaths(appDir) {
   Object.defineProperty(configurePaths, 'sourceWeb', _getValueObj(conf.source.web, _defaultPaths.sourceWeb));
   Object.defineProperty(configurePaths, 'sourceHybrid', _getValueObj(conf.source.hybrid, _defaultPaths.sourceHybrid));
   Object.defineProperty(configurePaths, 'sourceJavascript', _getValueObj(conf.source.javascript, _defaultPaths.sourceJavascript));
+  Object.defineProperty(configurePaths, 'sourceTests', _getValueObj(conf.source.tests, _defaultPaths.sourceTests));
   Object.defineProperty(configurePaths, 'sourceThemes', _getValueObj(conf.source.themes, _defaultPaths.sourceThemes));
   Object.defineProperty(configurePaths, 'stagingHybrid', _getValueObj(conf.staging.hybrid, _defaultPaths.stagingHybrid));
   Object.defineProperty(configurePaths, 'stagingWeb', _getValueObj(conf.staging.web, _defaultPaths.stagingWeb));
@@ -47,6 +49,7 @@ function _getDefaultPaths() {
   Object.defineProperty(defaultPaths, 'sourceWeb', _getValueObj('src-web'));
   Object.defineProperty(defaultPaths, 'sourceHybrid', _getValueObj('src-hybrid'));
   Object.defineProperty(defaultPaths, 'sourceJavascript', _getValueObj('js'));
+  Object.defineProperty(defaultPaths, 'sourceTests', _getValueObj('tests'));
   Object.defineProperty(defaultPaths, 'sourceThemes', _getValueObj('themes'));
   Object.defineProperty(defaultPaths, 'stagingHybrid', _getValueObj('hybrid'));
   Object.defineProperty(defaultPaths, 'stagingWeb', _getValueObj('web'));