+++ RELEASENOTES.md
@@ -1,5 +1,9 @@
 ## Release Notes for generator-oraclejet ##
 
+### 4.0.0
+* Moved module into @oracle scope, changing the name to @oracle/generator-oraclejet
+* This module should no longer be installed directly.  Use @oracle/ojet-cli instead for scaffolding JET apps.
+
 ### 3.2.0
 * No changes